backend/requirements.txt
pyJWT
fastapi
pydantic
bcrypt
uvicorn[standard]
motor
ldap3
python-multipart